[CoreBundle, OrderBundle, Order] fix for case #478: use documents in correct state for sending via mail
